automotive-genai-app/requirements.txt
Flask==2.3.3
Flask-Cors==3.1.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.0.0
Pillow==10.0.0
numpy==1.24.3
gunicorn==21.2.0
